--- src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/AgreedDraftOrderCollection.java
@@ -1,19 +1,0 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class AgreedDraftOrderCollection implements HasCaseDocument {
-    private AgreedDraftOrder value;
-
-}
