--- src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/draftorders/upload/DraftOrder.java
@@ -1,33 +1,0 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.upload;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.HasCaseDocument;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.OrderStatus;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.OrderToBeReviewed;
-
-import java.time.LocalDate;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-@Builder
-@AllArgsConstructor
-@EqualsAndHashCode
-public class DraftOrder implements HasCaseDocument, OrderToBeReviewed {
-
-    private OrderStatus status;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
-    private LocalDate submittedDate;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
-    private LocalDate reviewedDate;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
-    private LocalDate notificationSentDate;
-
-}
