+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/draftorders/upload/suggested/SuggestedPensionSharingAnnexCollection.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.upload;
+package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.upload.suggested;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
@@ -12,8 +12,7 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.HasCaseDocument;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class SuggestedDraftOrderCollection implements HasCaseDocument {
+public class SuggestedPensionSharingAnnexCollection implements HasCaseDocument {
 
-    private SuggestedDraftOrder value;
-
+    private SuggestedPensionSharingAnnex value;
 }
